Tweak: Separate style replacements
