+++ src/hoc/withHtmlAttributes.js
@@ -1,11 +1,11 @@
-import { useEffect, useMemo, useState } from '@wordpress/element';
+import { useEffect, useMemo } from '@wordpress/element';
 import { InspectorAdvancedControls } from '@wordpress/block-editor';
 import { TextControl } from '@wordpress/components';
+import { applyFilters } from '@wordpress/hooks';
 
 import { useUpdateEffect } from 'react-use';
 
 import { convertInlineStyleStringToObject } from '@utils/convertInlineStyleStringToObject';
-import { replaceTags } from '../dynamic-tags/utils';
 
 export const booleanAttributes = [
 	'allowfullscreen',
@@ -74,37 +74,12 @@ export function withHtmlAttributes( WrappedComponent ) {
 			align,
 		} = attributes;
 
-		const [ styleWithReplacements, setStyleWithReplacements ] = useState( '' );
 		const { style = '', href, ...otherAttributes } = htmlAttributes;
-
-		useEffect( () => {
-			async function getReplacements() {
-			// Check if any replacements need to be made if not, do nothing.
-				if ( ! style.includes( '{{' ) ) {
-					setStyleWithReplacements( style );
-					return;
-				}
-
-				const replacements = await replaceTags( style, context );
-
-				if ( ! replacements.length ) {
-					setStyleWithReplacements( style );
-					return;
-				}
-
-				const withReplacements = replacements.reduce( ( acc, { original, replacement, fallback } ) => {
-					if ( ! replacement ) {
-						return acc.replaceAll( original, fallback );
-					}
-
-					return acc.replaceAll( original, replacement );
-				}, style );
-
-				setStyleWithReplacements( withReplacements ? withReplacements : style );
-			}
-
-			getReplacements();
-		}, [ style, context ] );
+		const styleValue = applyFilters(
+			'generateblocks.editor.htmlAttributes.style',
+			style,
+			{ ...props }
+		);
 
 		useUpdateEffect( () => {
 			const layoutClasses = [ 'alignwide', 'alignfull' ];
@@ -120,8 +95,8 @@ export function withHtmlAttributes( WrappedComponent ) {
 			setAttributes( { className: newClasses.join( ' ' ) } );
 		}, [ align ] );
 
-		const inlineStyleObject = typeof styleWithReplacements === 'string'
-			? convertInlineStyleStringToObject( styleWithReplacements )
+		const inlineStyleObject = typeof styleValue === 'string'
+			? convertInlineStyleStringToObject( styleValue )
 			: '';
 		const combinedAttributes = {
 			...otherAttributes,
